requirements.txt
streamlit
openai
python-dotenv
httpx
requests
